# api/openapi.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/api/items": {"get": {"tags": ["items"], "summary": "Get Items", "operationId": "get_items", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Items-Get Items", "type": "array", "items": {"$ref": "#/components/schemas/Items"}}}}}}}, "post": {"tags": ["items"], "summary": "Create Item", "operationId": "create_item", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Items"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponseMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/python": {"get": {"summary": "Get Hello", "operationId": "get_hello", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response -Get Hello", "type": "string"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Items": {"title": "Items", "required": ["name", "price"], "type": "object", "properties": {"id": {"title": "Id", "type": "string"}, "created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "name": {"title": "Name", "type": "string"}, "price": {"title": "Price", "type": "integer"}}, "description": "Represents a Items record"}, "ResponseMessage": {"title": "ResponseMessage", "required": ["message"], "type": "object", "properties": {"message": {"title": "Message", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}