# openAPI/models/Items.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Represents a Items record
 */
export type Items = {
    id?: string;
    created_at?: string;
    name: string;
    price: number;
};